python/requirements.txt
mongoengine
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo
pydantic
uvicorn
typing
datetime
requests
beautifulsoup4
lxml
pandas
numpy
python-dotenv==1.0.0
aiohttp
geopy
bs4
openai>=1.0.0
rich
mongoengine
pymongo
